+++ backend/Squirrel.ConsoleApp/appsettings.json
@@ -10,6 +10,6 @@
     "BaseUrl": "http://localhost:44567"
   },
   "SignalRSettings": {
-    "HubConnectionUrl": "https://localhost:7244/SquirrelHub"
+    "HubConnectionUrl": "https://localhost:7244/ConsoleAppHub"
   }
 }
